+++ src/main/kotlin/io/snyk/plugin/settings/SnykProjectSettingsConfigurable.kt
@@ -20,8 +20,8 @@ import io.snyk.plugin.isUrlValid
 import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.ui.SnykBalloonNotificationHelper
 import io.snyk.plugin.ui.SnykSettingsDialog
-import snyk.common.lsp.settings.FolderConfigSettings
 import snyk.common.lsp.LanguageServerWrapper
+import snyk.common.lsp.settings.FolderConfigSettings
 import javax.swing.JComponent
 
 class SnykProjectSettingsConfigurable(
@@ -126,7 +126,7 @@ class SnykProjectSettingsConfigurable(
             }
 
             languageServerWrapper.refreshFeatureFlags()
-            languageServerWrapper.updateConfiguration()
+            languageServerWrapper.updateConfiguration(true)
         }
 
         if (rescanNeeded) {