+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added a description of custom endpoints to settings dialog.
 - Add option to ignore IaC issues
 ### Fixed
+- only ask to scan folders that are known to language server
 - folder-specific configs are availabe on opening projects, not only on restart of the IDE
 - display open source issues in Rider. Previously, as the project.assets.json is in a derived folder, it was filtered.
 - correctly display and update base branch name for Net New Issues